Updated FCC lattice and tests
I updated make_fcc_v3 as well as make_all_mirrors and make_mirror_3D. I also made Check_FCC_Functionality_v1 for checking the FCC. This file has notes on the limitations on possible inputs for bounds, etc.. that can be used. You should be able to use these results to fix any problem you are having with the fcc in your code. It is possible something broke when I tried to make the bounds more general (forgetting why I wasnt able to do that in the first place..)