Radius dependent Random Move and Move Plots
Updated the random move step to move particles proportional to the particle radius. I also added 2 extra plots that show the total distance a particle has moved and the final distance the particle is located with respect to its starting position. This will be helpful for setting giggles later.  The current code has a FF error and the [0,0,0] cord is violated.