Added checks into functions and verbose outputs
I added a check function that verifies [0,0,0] particle exists, no particles are repeated, and that particles do not overlap. This check function is called in each major function to ensure steps do not break these assumptions. If assumptions are broken a FLAG is thrown and the program goes into debug inside the problematic function.

NOTE: I also fixed the random delete particle step. This step could technically remove the center particle (with low probability as number of particles increased). This is why I think running the all verifications is better than a single stress test.